autentificacion en rutas iinfoContact

+++ api/controllers/infoContact.js
@@ -1,5 +1,7 @@
+const { restart } = require('nodemon')
 const InfoContact = require('../models/infoContact')
 const User = require('../models/user')
+const { ROWLOCK } = require('sequelize/types/table-hints')
 
 async function getAllinfoContact(req, res) {
 	try {
@@ -17,6 +19,11 @@ async function getAllinfoContact(req, res) {
 
 async function getOneInfo(req, res) {
 
+	const user = res.local.user
+	if (user.rol !== "admin" && user.id !==  req.params.id ){
+		return res.status(403).send("No estas autorizado") 
+	}
+
 	try {
 		const infoUser = await InfoContact.findByPk(req.params.id)
 		if (infoUser) {
@@ -30,25 +37,35 @@ async function getOneInfo(req, res) {
 }
 
 async function createInfoContactFunction(req, res) {
-    try {
    
+	const user = res.local.user
+	if (user.rol !== "admin" && user.id !==  req.params.id ){
+		return res.status(403).send("No estas autorizado") 
+	}
+	
+	try {
         const user = await User.findByPk(req.params.id);
         if (!user) {
             return res.status(404).json({ message: 'Usuario no encontrado' });
         }
         const infoContact = await InfoContact.create(req.body);
 		const userContact = await infoContact.setUser(user)
-        return res.status(201).json({ message: 'Información de contacto creada correctamente', infoContact });
+        return res.status(201).json({ message: 'Información de contacto creada correctamente', infoContact});
 
     } catch (error) {
 
         console.error('Error al crear información de contacto:', error);
         return res.status(500).json({ message: 'Error al crear información de contacto'});
     }
-
 }
 
 async function updateInfoContact(req, res) {
+	
+	const user = res.local.user
+	if (user.rol !== "admin" && user.id !==  req.params.id ){
+		return res.status(403).send("No estas autorizado") 
+	}
+	
 	try {
 		const [infoExist, info] = await InfoContact.update(req.body, {
 			returning: true,
@@ -67,6 +84,11 @@ async function updateInfoContact(req, res) {
 }
 
 async function deleteInfoContact(req, res) {
+	const user = res.local.user
+	if (user.rol !== "admin" && user.id !==  req.params.id ){
+		return res.status(403).send("No estas autorizado") 
+	}
+
 	try {
 		const user = await InfoContact.destroy({
 			where: {
